requirements.txt
torch==1.9.0
torchvision==0.10.0
PyQt5==5.15.4
Pillow==8.3.1
numpy==1.21.2